add trip pont presenter
